Struct lemipc::board::Map

pub struct Map {
    // some fields omitted
}

The Map structure defines the grid and the score.

Methods

impl Map

fn get<'a>(&'a self, x: usize, y: usize) -> Option<(i32, bool)>

The get function returns the (pid, team) cell.

fn get_pid<'a>(&'a self, x: usize, y: usize) -> Option<i32>

The get_pid function returns the pid cell.

fn get_score<'a>(&'a self) -> [usize; 2]

The get_score function returns the score.

fn get_turn<'a>(&'a self) -> i32

The get_turn function returns who must play.

fn set<'a>(&'a mut self, x: usize, y: usize, pid: i32, team: bool) -> bool

The set function writes the value at [X; Y] cell.

fn unset<'a>(&'a mut self, x: usize, y: usize) -> bool

The unset function writes the value at [X; Y] cell.

fn play_pawn<'a>(&'a mut self, pid: i32, compass: Compass) -> Result<bool, &'static str>

The play_pawn function moves the pawn.

fn spawn_pawn<'a>(&'a mut self, pid: i32) -> bool

The spawn_pawn function spawns a new pawn.

fn dead_pawn<'a>(&'a mut self, pid: i32) -> bool

The dead_pawn function removes a pawn and updates the turn of next pawn.

fn count_enemy<'a>(&'a self, pid: i32) -> Option<usize>

The count_enemy function returns the number of enemy around one pawn.

fn found_team<'a>(&'a self, pid: i32) -> Option<bool>

The found_team returns the team of pawn according to pid.

fn len_pawn<'a>(&'a self) -> usize

The len_pawn counts the number of pawn.

fn put_grid_team<'a>(&'a self, pid: i32)

The put_grid_team prints the grid with a limited vision of team.