Module lemipc::command

Reexports

pub use ffi;

Enums

Compass
IPC - compass help

Functions

cheat
The cheat function prints all the map.

email
The email function sends a message.

help
The help function prints all commands of game.

map
The map function prints the map according to the team of pid.

play
The play function checks if the user can play, if yes the pawn is moved.

quit
Dead dead dead !

receive
The receive function takes and prints the last message.

score
The score function prints the result of team.

start
The start function prints the first message of game.

turn
The turn function informs the user who must play.

whoiam
The whoiam function prints the pid.